rename Resources to Why Vacate and update imports
