+++ src/App.js
@@ -12,7 +12,7 @@ import ContactPage from "./Pages/ContactPage";
 import Footer from "./Components/Footer";
 import GetInvolvedPage from "./Pages/GetInvolvedPage";
 import GetStartedPage from "./Pages/GetStartedPage";
-import ResourcesPage from "./Pages/ResourcesPage";
+import WhyVacatePage from "./Pages/WhyVacatePage";
 import ContactSuccessPage from "./Pages/ContactSuccessPage";
 import OnboardingHome from "./Pages/Onboarding/OnboardingHome";
 import OnboardingMod1 from "./Pages/Onboarding/OnboardingMod1";
@@ -64,8 +64,8 @@ function App() {
                             <Route path="/get-started">
                                 <GetStartedPage />
                             </Route>
-                            <Route path="/resources">
-                                <ResourcesPage />
+                            <Route path="/why-vacate">
+                                <WhyVacatePage />
                             </Route>
 
                             {/* Volunteer onboarding pages */}
